requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
coverage==7.3.0
cryptography==41.0.7
Deprecated==1.2.14
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.3.1
django-extensions==3.2.3
django-safedelete==1.3.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gcloud==0.18.3
googleapis-common-protos==1.62.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
jwcrypto==1.5.1
oauth2client==4.1.3
packaging==23.1
pillow==10.2.0
protobuf==4.25.1
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.1
PyJWT==2.8.0
pyparsing==3.1.1
Pyrebase4==4.7.1
pyscopg2==66.0.2
python-decouple==3.8
python-dotenv==1.0.0
python-jwt==4.1.0
pytz==2023.3
PyYAML==6.0.1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.18
wrapt==1.16.0
